chore: setup blueprints folder to organize endpoint files
